menambahkan code update pada bagian admin

+++ admin/fungsi/fungsi.php
@@ -83,28 +83,10 @@ function edit_admin()
 
 	$hapus_foto = $r['foto'];
 
-		if(isset($_POST['ubahfoto']))
+	if(isset($_POST['ubahfoto']))
 	{
-		if ($row['foto']=="") 
+		if ($r['foto']=="") 
 		{
-				if ($foto != "") {
-				$allowed_ext = array('png','jpg');
-				$x = explode(".", $foto);
-				$ekstensi = strtolower(end($x));
-				$file_tmp = $_FILES['foto']['tmp_name'];
-				$angka_acak = rand(1,999);
-		   		$nama_file_baru = $angka_acak.'-'.$foto;
-		   		    if (in_array($ekstensi, $allowed_ext) === true) {
-		      			move_uploaded_file($file_tmp, 'img/'.$nama_file_baru);
-		      			$result =  mysqli_query($koneksi, "UPDATE tb_admin SET username='$username', password='$password', nama='$nama', kontak='$kontak', foto='$nama_file_baru' WHERE id='$id'");
-		      			
-		      			
-		    }
-
-
-
-			}
-		}else if ($row['foto']!="") {
 			if ($foto != "") {
 				$allowed_ext = array('png','jpg');
 				$x = explode(".", $foto);
@@ -112,18 +94,26 @@ function edit_admin()
 				$file_tmp = $_FILES['foto']['tmp_name'];
 				$angka_acak = rand(1,999);
 		   		$nama_file_baru = $angka_acak.'-'.$foto;
-		   		    if (in_array($ekstensi, $allowed_ext) === true) {
-		      			move_uploaded_file($file_tmp, 'img/'.$nama_file_baru);
-		      			$result =  mysqli_query($koneksi, "UPDATE tb_admin SET username='$username', password='$password', nama='$nama', kontak='$kontak', foto='$nama_file_baru' WHERE id='$id'");
-		      			if ($result) {
-		      				unlink("img/$hapus_foto");
-		      			}
-
-		      			
-		    }
-
-
-
+		   		if (in_array($ekstensi, $allowed_ext) === true) {
+		      		move_uploaded_file($file_tmp, 'img/'.$nama_file_baru);
+		      		$result =  mysqli_query($koneksi, "UPDATE tb_admin SET username='$username', password='$password', nama='$nama', kontak='$kontak', foto='$nama_file_baru' WHERE id='$id'");
+		   		}
+			}
+		} else if ($r['foto']!="") {
+			if ($foto != "") {
+				$allowed_ext = array('png','jpg');
+				$x = explode(".", $foto);
+				$ekstensi = strtolower(end($x));
+				$file_tmp = $_FILES['foto']['tmp_name'];
+				$angka_acak = rand(1,999);
+		   		$nama_file_baru = $angka_acak.'-'.$foto;
+		   		if (in_array($ekstensi, $allowed_ext) === true) {
+		      		move_uploaded_file($file_tmp, 'img/'.$nama_file_baru);
+		      		$result =  mysqli_query($koneksi, "UPDATE tb_admin SET username='$username', password='$password', nama='$nama', kontak='$kontak', foto='$nama_file_baru' WHERE id='$id'");
+		      		if ($result) {
+		      			unlink("img/$hapus_foto");
+		      		}
+		   		}
 			}
 		}	
 	}
@@ -131,8 +121,6 @@ function edit_admin()
 	if (empty($_POST['foto'])) {
 		return  mysqli_query($koneksi, "UPDATE tb_admin SET username='$username', password='$password', nama='$nama', kontak='$kontak' WHERE id='$id'");
 	}
-
-
 }
 
 // ---------------------------------------------KARYAWAN SECTION------------------------------------------------------------
@@ -150,7 +138,7 @@ function simpan_karyawan()
 	$kontak = $_POST['kontak'];
 	$foto = $_FILES['foto']['name'];
 
-if ($foto!= "") {
+	if ($foto!= "") {
 		$allowed_ext = array('png','jpg');
 		$x = explode(".", $foto);
 		$ext = strtolower(end($x));
@@ -160,7 +148,6 @@ if ($foto!= "") {
 		if (in_array($ext, $allowed_ext)===true) {
 			move_uploaded_file($file_tmp, 'img/karyawan/'.$nama_file_baru);
 			$res = mysqli_query($koneksi, "INSERT INTO tb_karyawan SET nip='$nip', username='$username', password='$password', nama='$nama', tempat_lahir='$tempat_lahir', tanggal_lahir='$tanggal_lahir', alamat='$alamat', kontak='$kontak', foto='$nama_file_baru'");
-
 		}
 	}else if (empty($_POST['foto'])) {
 		$res = mysqli_query($koneksi, "INSERT INTO tb_karyawan SET nip='$nip', username='$username', password='$password', nama='$nama', tempat_lahir='$tempat_lahir', tanggal_lahir='$tanggal_lahir', alamat='$alamat', kontak='$kontak'");
@@ -198,8 +185,68 @@ function select_karyawan_2()
 	$select = mysqli_query($koneksi, "SELECT count(id) AS jkaryawan FROM tb_karyawan");
 	$r = mysqli_fetch_array($select);
 	echo $r['jkaryawan'];
-
 }
+
+function update_karyawan()
+{
+	global $koneksi;
+	$id = $_POST['id'];
+	$nip = $_POST['nip'];
+	$username = $_POST['username'];
+	$password = md5($_POST['password']);
+	$nama = $_POST['nama'];
+	$tempat_lahir = $_POST['tempat_lahir'];
+	$tanggal_lahir = $_POST['tanggal_lahir'];
+	$alamat = $_POST['alamat'];
+	$kontak = $_POST['kontak'];
+	$foto = $_FILES['foto']['name'];
+
+	// unlink 
+	$sql = mysqli_query($koneksi, "SELECT * FROM tb_karyawan WHERE id='$id'");
+	$r = mysqli_fetch_array($sql);
+
+	$hapus_foto = $r['foto'];
+
+	if(isset($_POST['ubahfoto']))
+	{
+		if ($r['foto']=="") 
+		{
+			if ($foto != "") {
+				$allowed_ext = array('png','jpg');
+				$x = explode(".", $foto);
+				$ekstensi = strtolower(end($x));
+				$file_tmp = $_FILES['foto']['tmp_name'];
+				$angka_acak = rand(1,999);
+		   		$nama_file_baru = $angka_acak.'-'.$foto;
+		   		if (in_array($ekstensi, $allowed_ext) === true) {
+		      		move_uploaded_file($file_tmp, 'img/karyawan/'.$nama_file_baru);
+		      		$result =  mysqli_query($koneksi, "UPDATE tb_karyawan SET nip='$nip', username='$username', password='$password', nama='$nama', tempat_lahir='$tempat_lahir', tanggal_lahir='$tanggal_lahir', alamat='$alamat', kontak='$kontak', foto='$nama_file_baru' WHERE id='$id'");
+		   		}
+			}
+		} else if ($r['foto']!="") {
+			if ($foto != "") {
+				$allowed_ext = array('png','jpg');
+				$x = explode(".", $foto);
+				$ekstensi = strtolower(end($x));
+				$file_tmp = $_FILES['foto']['tmp_name'];
+				$angka_acak = rand(1,999);
+		   		$nama_file_baru = $angka_acak.'-'.$foto;
+		   		if (in_array($ekstensi, $allowed_ext) === true) {
+		      		move_uploaded_file($file_tmp, 'img/karyawan/'.$nama_file_baru);
+		      		$result =  mysqli_query($koneksi, "UPDATE tb_karyawan SET nip='$nip', username='$username', password='$password', nama='$nama', tempat_lahir='$tempat_lahir', tanggal_lahir='$tanggal_lahir', alamat='$alamat', kontak='$kontak', foto='$nama_file_baru' WHERE id='$id'");
+		      		if ($result) {
+		      			unlink("img/karyawan/$hapus_foto");
+		      		}
+		   		}
+			}
+		}	
+	}
+
+	if (empty($_POST['foto'])) {
+		return  mysqli_query($koneksi, "UPDATE tb_karyawan SET nip='$nip', username='$username', password='$password', nama='$nama', tempat_lahir='$tempat_lahir', tanggal_lahir='$tanggal_lahir', alamat='$alamat', kontak='$kontak' WHERE id='$id'");
+	}
+}
+
 // ----------------------------------------SETTING SECTION-------------------------------------------------------------------------------\\
 function ubah_jam_masuk()
 {
@@ -257,14 +304,12 @@ function hapus_keterangan()
 	if ($delete) {
 		echo '<script>alert("data sudah dihapus!")</script>';
 	}
-
-
 }
+
 // ----------------------------------------FUNCTION URL, KEEP IT BELOW!!------------------------------------------------------------------
 function url()
 {
 	return $url = "//localhost/aplikasiKepegawaian/assets/";
 
 }
-
- ?>
+?>
